pyproject.toml
[project]
name = "taskbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0.0",
    "fastapi>=0.68.0",
    "ffmpeg-python>=0.2.0",
    "jinja2>=3.0.0",
    "nltk>=3.8.1",
    "pandas>=1.3.0",
    "plotly>=5.3.0",
    "pydub>=0.25.1",
    "python-dateutil>=2.8.2",
    "python-dotenv>=0.19.0",
    "pytz>=2023.3",
    "spacy>=3.0.0",
    "speechrecognition>=3.8.1",
    "sqlalchemy>=1.4.0",
    "uvicorn>=0.15.0",
]
